+++ c/dsls/lemon_assert.h
@@ -1,0 +1,11 @@
+#ifndef LEMON_ASSERT_H
+#define LEMON_ASSERT_H
+
+void lemon_assert();
+#ifndef NDEBUG
+#  define assert(X) if(!(X))lemon_assert(__FILE__,__LINE__)
+#else
+#  define assert(X)
+#endif
+
+#endif // LEMON_ASSERT_H
