+++ c/dsls/lemon_error.h
@@ -1,0 +1,7 @@
+#ifndef LEMON_ERROR_H
+#define LEMON_ERROR_H
+#include "stdarg.h"
+
+void ErrorMsg(const char *, int,const char *, ...);
+
+#endif // LEMON_ERROR_H
